# PERF: DataFrame.join with how=left|right and sort=True 